+++ src/main/java/com/wolfbe/distributedidsdk/PressTestSync.java
@@ -54,6 +54,6 @@ public class PressTestSync {
         }
 
         logger.info("invokeSync test all num is: {}, all cast time: {} millsec, all throughput: {} send/millsec",
-                amount, allcast, amount/allcast);
+                amount, allcast, (double)amount/allcast);
     }
 }
